+++ services/service-warranty/src/main/java/com/example/service_warranty/dto/WarrantyValidationDto.java
@@ -12,7 +12,7 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 public class WarrantyValidationDto {
-    private Long warrantyRequestId;
+    private Integer warrantyRequestId;
     private Boolean isValid;
     private String validationReason;
     private String validatedBy;
@@ -23,7 +23,7 @@ public class WarrantyValidationDto {
     @NoArgsConstructor
     @AllArgsConstructor
     public static class ConditionResultDto {
-        private Long conditionId;
+        private Integer conditionId;
         private Boolean passed;
         private String notes;
     }
